requirements.txt
arabic-reshaper==3.0.0
astroid==4.0.1
blinker==1.9.0
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
dill==0.4.0
Flask==3.1.3
fpdf==1.7.2
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
pillow==11.3.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
pyfiglet==1.0.4
pylint==4.0.0
python-bidi==0.6.6
reportlab==4.4.3
termcolor==3.1.0
tomlkit==0.13.3
Werkzeug==3.1.6
